implement sqlite search (but can't test yet)

+++ src/download.rs
@@ -1,12 +1,14 @@
 use std::io::Write;
 use std::fs::File;
-use std::path::Path;
+use std::path::PathBuf;
 use thiserror::Error;
 
 use crate::log;
 
 #[derive(Error, Debug)]
 pub enum DownloadError {
+    #[error("Local error: {0}")]
+    LocalError(String),
     #[error("I/O error: {0}")]
     Io(#[from] std::io::Error),
     #[error("HTTP request error: {0}")]
@@ -22,15 +24,29 @@ pub enum DownloadError {
     // Ok(())
 //}
 
+fn db_dir() -> PathBuf {
+    PathBuf::from("/storage/emulated/0/Documents/")
+}
+
+fn db_fname() -> &'static str {
+    "kvideomanager.sqlite"
+}
+
+pub fn db_full_path() -> PathBuf {
+    db_dir().join(db_fname())
+}
+
 pub async fn download_db() -> Result<(), DownloadError> {
-    let tmp_dir = Path::new("/storage/emulated/0/Documents/");
-    log!("tmp_dir: '{}'", tmp_dir.display());
-    let target = "http://www.davidfaure.fr/kvideomanager/kvideomanager.sqlite";
-    let response = reqwest::get(target).await?;
+    let target_dir = db_dir();
+    if !target_dir.exists() {
+        let error_msg = format!("Local dir does not exist: {}", target_dir.display());
+        return Err(DownloadError::LocalError(error_msg));
+    }
+    let url = "http://www.davidfaure.fr/kvideomanager/kvideomanager.sqlite";
+    let response = reqwest::get(url).await?;
 
     let mut dest = {
-        let fname = "kvideomanager.sqlite";
-        let file_path = tmp_dir.join(fname);
+        let file_path = db_full_path();
         log!("will be located under: '{:?}'", file_path);
         File::create(file_path)?
     };
